Update main.py
Deleting old images when a new one was posted was a crappy design; my intention was to eliminate the directory from filling up. If end users' devices didn't already cache the image in the group, they would just see a grey box, however, because the image would no longer exist on the server. Now I have the program set to flush all the jpegs from /var/www/html (or whatever the server's directory is) when the directory contains more than 49 images.